Initial public commit
Names should be all gone from ticket testdata, and any 6-character
Amtrak reservation IDs are either changed to FAKE##, or they're in the
past, so they can't be used to get a ticket anymore.

Code is still embarrassing in many places, and littered with TODOs and
FIXMEs. But at least now I have a public README that I can point friends
to when they want to try it out. And I can use the Github issue tracker.