Set no-console in tslint
console.log logs to StackDriver Logging in Apps Script environments.
This particular app should generally avoid doing this, because I'd
rather not log a user's Amtrak reservation number - it's kind of
sensitive.

And yeah, the spec/testdata directory has a bunch of my real
reservations numbers, but they're for trains in the past, so I'm not
concerned about someone stealing those reservations.