Consider changed reservations when syncing calendar events
This rewrites AutoSync to use a ReservationCollection. The resulting
code is much simpler. Calendar events for a reservation are synced with
logic in Reservation.ts, where we add trains that don't have calendar
events, and remove calendar events that don't have trains.

This also adds a clearGlitchedEvents() method, which removes any
calendar event containing the strings "Amtrak2Calendar" and "undefined",
because they're clearly from Javascript errors. These events shouldn't
ever be created, but if they are, they'll get cleaned up.