Ignore reservation emails without an attachment
Fixes #1.

At least, I think it does. I don't have an email message that will
reproduce the failure described in the reported issue, but from tracing
the code, it looks like the "undefined" came from an out-of-bounds array
access. We now check for length of attachments(), i.e., we check that
the message has an attachment at all. If not, we skip it.